:art: Edit id field of booth model
Booth 모델의 id 필드를 bigint에서 int로 바꿨습니다.

+++ booth/models.py
@@ -4,7 +4,7 @@ from core.models import Base_image
 # Create your models here.
 # 부스
 class Booth(models.Model):
-    id = models.BigIntegerField(primary_key=True)
+    id = models.IntegerField(primary_key=True)
     name = models.CharField(max_length=255)
     TYPE_CHOICES=(
         # 자유롭게 수정
